sanchez_polo_drowsy/requirements.txt
opencv-python
dlib
imutils
scipy
pygame
PyYAML
stomp.py==7.0.0
websocket-client
requests
